Reorganize bamboo module
