+++ src/java/growthcraft/bamboo/BambooFuelHandler.java
@@ -1,8 +1,8 @@
 package growthcraft.bamboo;
 
+import cpw.mods.fml.common.IFuelHandler;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import cpw.mods.fml.common.IFuelHandler;
 
 public class BambooFuelHandler implements IFuelHandler
 {
