+++ src/java/growthcraft/bamboo/ClientProxy.java
@@ -1,5 +1,12 @@
 package growthcraft.bamboo;
 
+import growthcraft.bamboo.entity.EntityBambooRaft;
+import growthcraft.bamboo.renderer.RenderBamboo;
+import growthcraft.bamboo.renderer.RenderBambooFence;
+import growthcraft.bamboo.renderer.RenderBambooWall;
+import growthcraft.bamboo.renderer.RenderBambooScaffold;
+import growthcraft.bamboo.renderer.RenderBambooRaft;
+
 import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class ClientProxy extends CommonProxy
