+++ src/java/growthcraft/bamboo/GrowthCraftBamboo.java
@@ -4,6 +4,38 @@ import java.io.File;
 import java.util.HashMap;
 import java.util.List;
 
+import growthcraft.bamboo.block.BlockBamboo;
+import growthcraft.bamboo.block.BlockBambooDoor;
+import growthcraft.bamboo.block.BlockBambooFence;
+import growthcraft.bamboo.block.BlockBambooFenceGate;
+import growthcraft.bamboo.block.BlockBambooLeaves;
+import growthcraft.bamboo.block.BlockBambooScaffold;
+import growthcraft.bamboo.block.BlockBambooShoot;
+import growthcraft.bamboo.block.BlockBambooSlab;
+import growthcraft.bamboo.block.BlockBambooStairs;
+import growthcraft.bamboo.block.BlockBambooStalk;
+import growthcraft.bamboo.block.BlockBambooWall;
+import growthcraft.bamboo.entity.EntityBambooRaft;
+import growthcraft.bamboo.event.BonemealEventBamboo;
+import growthcraft.bamboo.item.ItemBamboo;
+import growthcraft.bamboo.item.ItemBambooCoal;
+import growthcraft.bamboo.item.ItemBambooDoor;
+import growthcraft.bamboo.item.ItemBambooRaft;
+import growthcraft.bamboo.item.ItemBambooShoot;
+import growthcraft.bamboo.item.ItemBambooSlab;
+import growthcraft.bamboo.world.BiomeGenBamboo;
+import growthcraft.bamboo.world.WorldGenBamboo;
+import growthcraft.bamboo.world.WorldGeneratorBamboo;
+
+import cpw.mods.fml.common.Mod;
+import cpw.mods.fml.common.Mod.EventHandler;
+import cpw.mods.fml.common.Mod.Instance;
+import cpw.mods.fml.common.SidedProxy;
+import cpw.mods.fml.common.event.FMLInitializationEvent;
+import cpw.mods.fml.common.event.FMLPostInitializationEvent;
+import cpw.mods.fml.common.event.FMLPreInitializationEvent;
+import cpw.mods.fml.common.registry.EntityRegistry;
+import cpw.mods.fml.common.registry.GameRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockSlab;
 import net.minecraft.init.Blocks;
@@ -20,17 +52,6 @@ import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.common.config.Configuration;
 import net.minecraftforge.common.config.Property;
 import net.minecraftforge.oredict.OreDictionary;
-import cpw.mods.fml.common.Mod;
-import cpw.mods.fml.common.Mod.EventHandler;
-import cpw.mods.fml.common.Mod.Instance;
-import cpw.mods.fml.common.SidedProxy;
-import cpw.mods.fml.common.event.FMLInitializationEvent;
-import cpw.mods.fml.common.event.FMLPostInitializationEvent;
-import cpw.mods.fml.common.event.FMLPreInitializationEvent;
-import cpw.mods.fml.common.registry.EntityRegistry;
-import cpw.mods.fml.common.registry.GameRegistry;
-
-@Mod(modid = "Growthcraft|Bamboo",name = "Growthcraft Bamboo",version = "2.1.1",dependencies = "required-after:Growthcraft")
 
 public class GrowthCraftBamboo
 {
