+++ src/java/growthcraft/bamboo/block/BlockBamboo.java
@@ -1,11 +1,12 @@
-package growthcraft.bamboo;
+package growthcraft.bamboo.block;
 
 import growthcraft.core.GrowthCraftCore;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.client.renderer.texture.IIconRegister;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class BlockBamboo extends Block
 {
