+++ src/java/growthcraft/bamboo/block/BlockBambooFenceGate.java
@@ -1,11 +1,13 @@
-package growthcraft.bamboo;
+package growthcraft.bamboo.block;
 
 import growthcraft.core.GrowthCraftCore;
+import growthcraft.bamboo.GrowthCraftBamboo;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.BlockFenceGate;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.util.IIcon;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class BlockBambooFenceGate extends BlockFenceGate
 {
