+++ src/java/growthcraft/bamboo/block/BlockBambooLeaves.java
@@ -1,10 +1,12 @@
-package growthcraft.bamboo;
-
-import growthcraft.core.GrowthCraftCore;
+package growthcraft.bamboo.block;
 
 import java.util.ArrayList;
 import java.util.Random;
 
+import growthcraft.core.GrowthCraftCore;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockLeavesBase;
 import net.minecraft.block.material.Material;
@@ -17,8 +19,6 @@ import net.minecraft.world.ColorizerFoliage;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.common.IShearable;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class BlockBambooLeaves extends BlockLeavesBase implements IShearable
 {
