+++ src/java/growthcraft/bamboo/block/BlockBambooScaffold.java
@@ -1,9 +1,12 @@
-package growthcraft.bamboo;
-
-import growthcraft.core.GrowthCraftCore;
+package growthcraft.bamboo.block;
 
 import java.util.Random;
 
+import growthcraft.bamboo.renderer.RenderBambooScaffold;
+import growthcraft.core.GrowthCraftCore;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.client.renderer.texture.IIconRegister;
@@ -17,8 +20,6 @@ import net.minecraft.util.IIcon;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.common.util.ForgeDirection;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class BlockBambooScaffold extends Block
 {
