+++ src/java/growthcraft/bamboo/entity/EntityBambooRaft.java
@@ -1,7 +1,11 @@
-package growthcraft.bamboo;
+package growthcraft.bamboo.entity;
 
 import java.util.List;
 
+import growthcraft.bamboo.GrowthCraftBamboo;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.entity.Entity;
@@ -15,8 +19,6 @@ import net.minecraft.util.AxisAlignedBB;
 import net.minecraft.util.DamageSource;
 import net.minecraft.util.MathHelper;
 import net.minecraft.world.World;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class EntityBambooRaft extends Entity
 {
