+++ src/java/growthcraft/bamboo/item/ItemBamboo.java
@@ -1,6 +1,10 @@
-package growthcraft.bamboo;
+package growthcraft.bamboo.item;
 
+import growthcraft.bamboo.GrowthCraftBamboo;
 import growthcraft.core.GrowthCraftCore;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.entity.Entity;
@@ -9,8 +13,6 @@ import net.minecraft.init.Blocks;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class ItemBamboo extends Item
 {
