+++ src/java/growthcraft/bamboo/item/ItemBambooCoal.java
@@ -1,10 +1,11 @@
-package growthcraft.bamboo;
+package growthcraft.bamboo.item;
 
 import growthcraft.core.GrowthCraftCore;
-import net.minecraft.client.renderer.texture.IIconRegister;
-import net.minecraft.item.Item;
+
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
+import net.minecraft.client.renderer.texture.IIconRegister;
+import net.minecraft.item.Item;
 
 public class ItemBambooCoal extends Item
 {
