+++ src/java/growthcraft/bamboo/item/ItemBambooDoor.java
@@ -1,6 +1,10 @@
-package growthcraft.bamboo;
+package growthcraft.bamboo.item;
 
+import growthcraft.bamboo.GrowthCraftBamboo;
 import growthcraft.core.GrowthCraftCore;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.entity.player.EntityPlayer;
@@ -8,8 +12,6 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.MathHelper;
 import net.minecraft.world.World;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class ItemBambooDoor extends Item
 {
