+++ src/java/growthcraft/bamboo/item/ItemBambooRaft.java
@@ -1,9 +1,12 @@
-package growthcraft.bamboo;
-
-import growthcraft.core.GrowthCraftCore;
+package growthcraft.bamboo.item;
 
 import java.util.List;
 
+import growthcraft.bamboo.entity.EntityBambooRaft;
+import growthcraft.core.GrowthCraftCore;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
@@ -15,8 +18,6 @@ import net.minecraft.util.MathHelper;
 import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.util.Vec3;
 import net.minecraft.world.World;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class ItemBambooRaft extends ItemBoat
 {
