+++ src/java/growthcraft/bamboo/model/ModelBambooRaft.java
@@ -1,12 +1,11 @@
-package growthcraft.bamboo;
+package growthcraft.bamboo.model;
 
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.client.model.ModelBase;
 import net.minecraft.client.model.ModelRenderer;
 import net.minecraft.entity.Entity;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
-@SideOnly(Side.CLIENT)
 public class ModelBambooRaft extends ModelBase
 {
 	public ModelRenderer box;
