+++ src/java/growthcraft/bamboo/renderer/RenderBamboo.java
@@ -1,6 +1,11 @@
-package growthcraft.bamboo;
+package growthcraft.bamboo.renderer;
 
+import growthcraft.bamboo.GrowthCraftBamboo;
+import growthcraft.bamboo.block.BlockBambooStalk;
 import growthcraft.core.Utils;
+
+import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
+import cpw.mods.fml.client.registry.RenderingRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockDoor;
 import net.minecraft.client.renderer.EntityRenderer;
@@ -9,11 +14,7 @@ import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.init.Blocks;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.IBlockAccess;
-
 import org.lwjgl.opengl.GL11;
-
-import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
-import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class RenderBamboo implements ISimpleBlockRenderingHandler
 {
