+++ src/java/growthcraft/bamboo/renderer/RenderBambooRaft.java
@@ -1,11 +1,14 @@
-package growthcraft.bamboo;
+package growthcraft.bamboo.renderer;
+
+import growthcraft.bamboo.GrowthCraftBamboo;
+import growthcraft.bamboo.entity.EntityBambooRaft;
+import growthcraft.bamboo.model.ModelBambooRaft;
 
 import net.minecraft.client.model.ModelBase;
 import net.minecraft.client.renderer.entity.Render;
 import net.minecraft.entity.Entity;
 import net.minecraft.util.MathHelper;
 import net.minecraft.util.ResourceLocation;
-
 import org.lwjgl.opengl.GL11;
 
 public class RenderBambooRaft extends Render
