+++ src/java/growthcraft/bamboo/renderer/RenderBambooScaffold.java
@@ -1,13 +1,16 @@
-package growthcraft.bamboo;
+package growthcraft.bamboo.renderer;
 
+import growthcraft.bamboo.GrowthCraftBamboo;
+import growthcraft.bamboo.block.BlockBambooScaffold;
 import growthcraft.core.Utils;
+
+import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
+import cpw.mods.fml.client.registry.RenderingRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.RenderBlocks;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.IBlockAccess;
-import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
-import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class RenderBambooScaffold implements ISimpleBlockRenderingHandler
 {
